requirements.txt
python-decouple~=3.8
Telethon~=1.30.3